instalando git, mariadb

+++ grupo4-deploy.sh
@@ -13,21 +13,21 @@ LYELLOW='\033[1;33m'
 
 
 
-echo  -e "\n${LRED} ########################################################################${NC}"
-echo -e "\n${LRED} ######################Inicio de Automatizacion#########################${NC}"
-echo -e "\n${LRED} ########################################################################${NC}"
+echo  -e "\n${LGREEN} ########################################################################${NC}"
+echo -e "\n${LGREEN} ######################Inicio de Automatizacion#########################${NC}"
+echo -e "\n${LGREEN} ########################################################################${NC}"
 
 
 
 #Usuario Root solamente
 if [ "${USERID}" -ne 0 ]; then
-       echo -e "\n${LRED} El usuario debe ser Root ${NC}"
+       echo -e "\n${LGREEN} El usuario debe ser Root ${NC}"
        exit
 fi
 
 
 
-echo -e "\n${LBLUE} Actualizando servidor ${NC}"
+echo -e "\n${LGREEN} Actualizando servidor ${NC}"
 apt-get update
 echo -e "\n${LGREEN} El Servidor se encuentra Actualizado ...${NC}"
 
@@ -44,24 +44,54 @@ echo -e "\n${LGREEN} El Servidor se encuentra Actualizado ...${NC}"
 
 
   if dpkg -s apache2 > /dev/null 2>&1; then
-        echo -e "\n${LBLUE} El Apache2 se encuentra ya instalado ...${NC}"
+        echo -e "\n${LGREEN} El Apache2 se encuentra ya instalado ...${NC}"
   else
-        echo -e "\n${LBLUE} Instalanco Apache2 ...${NC}"
+        echo -e "\n${LYELLOW} Instalanco Apache2 ...${NC}"
         apt install -y apache2
+        #php
         apt install -y php libapache2-mod-php php-mysql
         systemctl start apache2
         systemctl enable apache2
         mv /var/www/html/index.html /var/www/html/index.html.bkp
-        echo -e "\n${LBLUE} Apache2 instalado ...${NC}"
+        echo -e "\n${LGREEN} Apache2 instalado ...${NC}"}
+        echo -e "\n${LGREEN} Version php:${NC} php -version"
    fi
 
 
 
 #git
-apt install -y git
-echo -e "\n${LYELLOW}instalando GIT ...${NC}"
+if dpkg -s git > /dev/null 2>&1; then
+  echo -e "\n${LGREEN} El git se encuentra ya instalado ...${NC}"
+else
+  apt install -y git
+  echo -e "\n${LYELLOW}instalando GIT ...${NC}"
+if
 
-#php
+
+##################################STAGE 2: [Build]######################
+#Voy a verificar primeramente si existe el repositorio porque desde allí sacae la base de destinations
+#Luego instala Mariadb
+#el script debe permitir evaluar la existencia del proyecto, si existe un git pull y si no existe un git clone
+#Mover al directorio donde se guardar los archivos de configuración de apache /var/www/html/
+#Testear existencia del codigo de la aplicación
+#Se recomienda renombrar en index.html default de apache para que no pise con el index.php
+
+if [ -d "$repo" ]; then
+    echo "La carpeta $repo existe"
+    git pull
+else
+    echo -e "\n${LYELLOW} Installing web ...${NC}"
+    sleep 1
+    git clone -b app-295devops-travel https://github.com/betsyninoska/$repo.git
+    cp -r $repo/* /var/www/html
+fi
+
+
+
+#Ajustar el config de php para que soporte los archivos dinamicos de php agregando index.php
+#Testear la compatibilidad -> ejemplo http://localhost/info.php
+#Si te muestra resultado de una pantalla informativa php , estariamos funcional para la siguiente etapa.
+
 
 
 
@@ -77,38 +107,8 @@ else
     systemctl enable mariadb
 
     echo -e "\n${LBLUE}Configurando base de datos ...${NC}"
+
+    mariadb -h localhost  -u root -P 3306  -p devopstravel < app-295devops-travel/database/devopstravel.s
     #Base de datos (Se debe automatizar que se pueda agregar el pass de la base de datos al momento de ejecutar el script asi evitamos que datos sensibles queden en el repositorio)
 fi #cierre condicional
-
-
-
-##################################STAGE 2: [Build]######################
-
-#el script debe permitir evaluar la existencia del proyecto, si existe un git pull y si no existe un git clone
-
-#if [ -d "$repo" ]; then
-#    echo "La carpeta $repo existe"
-#    rm -rf $repo
-#fi
-
-#echo -e "\n\e[92mInstalling web ...\033[0m\n"
-#sleep 1
-#git clone -b app-295devops-travel https://github.com/betsyninoska/$repo.git
-#cp -r $repo/* /var/www/html
-
-#Mover al directorio donde se guardar los archivos de configuración de apache /var/www/html/
-#Testear existencia del codigo de la aplicación
-
-#Ajustar el config de php para que soporte los archivos dinamicos de php agregando index.php
-
-
-#Testear la compatibilidad -> ejemplo http://localhost/info.php
-#Si te muestra resultado de una pantalla informativa php , estariamos funcional para la siguiente etapa.
-
-
-
-
-#Se recomienda renombrar en index.html default de apache para que no pise con el index.php
-
-
 #Revisar la ingesta de datos en la base de datos
